Add `Cluster` info to `stream::Info``
