+++ async-nats/src/jetstream/consumer/mod.rs
@@ -23,6 +23,7 @@ use serde_json::json;
 use time::serde::rfc3339;
 
 use super::response::Response;
+use super::stream::ClusterInfo;
 use super::Context;
 use crate::jetstream::consumer;
 use crate::Error;
@@ -157,37 +158,6 @@ pub struct Info {
     pub push_bound: bool,
 }
 
-/// Information about the stream's, consumer's associated `JetStream` cluster
-#[derive(Debug, Default, Serialize, Deserialize, Clone, PartialEq, Eq)]
-pub struct ClusterInfo {
-    /// The cluster name.
-    #[serde(default)]
-    pub name: Option<String>,
-    /// The server name of the RAFT leader.
-    #[serde(default)]
-    pub leader: Option<String>,
-    /// The members of the RAFT cluster.
-    #[serde(default)]
-    pub replicas: Vec<PeerInfo>,
-}
-
-/// The members of the RAFT cluster
-#[derive(Debug, Default, Serialize, Deserialize, Clone, PartialEq, Eq)]
-pub struct PeerInfo {
-    /// The server name of the peer.
-    pub name: String,
-    /// Indicates if the server is up to date and synchronised.
-    pub current: bool,
-    /// Nanoseconds since this peer was last seen.
-    #[serde(with = "serde_nanos")]
-    pub active: Duration,
-    /// Indicates the node is considered offline by the group.
-    #[serde(default)]
-    pub offline: bool,
-    /// How many uncommitted operations this peer is behind the leader.
-    pub lag: Option<u64>,
-}
-
 /// Information about a consumer and the stream it is consuming
 #[derive(Debug, Default, Serialize, Deserialize, Clone, Copy, PartialEq, Eq)]
 pub struct SequencePair {
@@ -308,6 +278,10 @@ pub struct Config {
     /// Threshold for ephemeral consumer intactivity
     #[serde(default, with = "serde_nanos", skip_serializing_if = "is_default")]
     pub inactive_threshold: Duration,
+
+    /// Number of consumer replucas
+    #[serde(default, skip_serializing_if = "is_default")]
+    pub num_replicas: usize,
 }
 
 impl From<&Config> for Config {
