+++ async-nats/tests/jetstream_tests.rs
@@ -180,7 +180,11 @@ mod jetstream {
             .await
             .unwrap();
 
+        println!("STREAM INFO: {:#?}", stream.info);
         assert_eq!(stream.info.config.num_replicas, 3);
+        assert!(stream.info.cluster.is_some());
+        assert_eq!(stream.info.cluster.as_ref().unwrap().replicas.len(), 2);
+
         context.delete_stream("events2").await.unwrap();
     }
